refactor(fpss): extract accumulator.rs
Move OhlcvcAccumulator and the change_price_type helper out of fpss/mod.rs
into a dedicated accumulator submodule, along with their unit tests.
Accumulator fields and methods are exposed as pub(super) so decode_frame
(still in mod.rs for now) can keep updating them directly.
